switch to using a decorator
