+++ src/sentry/integrations/slack/__init__.py
@@ -1,10 +1,3 @@
-# handle additional_attachment_manager frist
-from .additional_attachment_manager import AdditionalAttachmentManager
-
-manager = AdditionalAttachmentManager()
-register_additional_attachment_generator = manager.register_additional_attachment_generator
-get_additional_attachment = manager.get_additional_attachment
-
 from sentry.rules import rules
 from sentry.utils.imports import import_submodules
